requirements.txt
# MoodTracker - AI Mental Health Platform Dependencies

# Core Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter>=3.5.0
flasgger>=0.9.7
Werkzeug>=3.0.0

# AI/ML Libraries
torch>=2.0.0
transformers>=4.30.0
tensorflow>=2.10.0
tensorflow-hub>=0.14.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.2.0

# Text Processing
tokenizers>=0.13.0
scipy>=1.9.0

# HTTP Requests & Authentication
requests>=2.28.0
python-dotenv>=1.0.0
google-auth>=2.0.0
PyJWT>=2.8.0
argon2-cffi>=23.1.0

# Data Processing
pytz>=2023.3
python-dateutil>=2.8.2

# Development & Testing
pytest>=7.0.0
pytest-flask>=1.2.0

# Production Server
gunicorn==21.2.0

# Phase 4: Redis & Caching
redis>=5.0.0
Flask-Caching>=2.1.0

# Phase 4: Monitoring & Error Tracking
sentry-sdk[flask]>=1.40.0

# Phase 4: Security Headers
Flask-Talisman>=1.1.0

# Phase 4: Testing
pytest-cov>=4.1.0
locust>=2.18.0

# Optional: Enhanced Performance
# accelerate>=0.21.0  # For faster transformers
# sentencepiece>=0.1.99  # For tokenization